backend/requirements.txt
requests
beautifulsoup4
feedparser
httpx
python-dotenv
apscheduler
rapidfuzz
pdfplumber
networkx
